requirements.txt
# Core LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
openai>=1.0.0

# RAG & Vector Store
pgvector>=0.3.0
psycopg2-binary>=2.9.0
langchain-postgres>=0.0.9

# Embeddings (本地CUDA加速，无需API费用)
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
torch>=2.0.0

# MCP Protocol
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0

# Document Loaders
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.15.0

# Web Search
ddgs>=0.1.0
requests>=2.32.0
beautifulsoup4>=4.12.0

# CLI
rich>=13.0.0
click>=8.1.0
prompt_toolkit>=3.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
httpx>=0.27.0
